requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
pyodbc>=5.0.0

# Kafka
aiokafka>=0.10.0

# Redis
redis>=5.0.0

# HTTP Client
httpx>=0.26.0

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
